Cleanup namespaces
