+++ library/Notifications/Widget/Calendar/Entry.php
@@ -4,83 +4,31 @@
 
 namespace Icinga\Module\Notifications\Widget\Calendar;
 
-use DateTime;
 use DateTimeInterface;
 use ipl\Html\Attributes;
 use ipl\Html\BaseHtmlElement;
 use ipl\Html\HtmlElement;
 use ipl\Html\Text;
-use ipl\I18n\Translation;
-use ipl\Web\Url;
-use ipl\Web\Widget\Link;
+use Icinga\Module\Notifications\Widget\TimeGrid;
 
 /**
- * An entry on a time grid
- *
- * @phpstan-type GridContinuationType self::FROM_PREV_GRID | self::TO_NEXT_GRID | self::ACROSS_GRID
- * @phpstan-type EdgeContinuationType self::ACROSS_LEFT_EDGE | self::ACROSS_RIGHT_EDGE | self::ACROSS_BOTH_EDGES
- * @phpstan-type ContinuationType GridContinuationType | EdgeContinuationType
+ * An entry on a calendar
  */
-class Entry extends BaseHtmlElement
+class Entry extends TimeGrid\Entry
 {
-    use Translation;
-
-    /** @var string Continuation of an entry that started on the previous grid */
-    public const FROM_PREV_GRID = 'from-prev-grid';
-
-    /** @var string Continuation of an entry that continues on the next grid */
-    public const TO_NEXT_GRID = 'to-next-grid';
-
-    /** @var string Continuation of an entry that started on the previous grid and continues on the next */
-    public const ACROSS_GRID = 'across-grid';
-
-    /** @var string Continuation of an entry that started on a previous grid row */
-    public const ACROSS_LEFT_EDGE = 'across-left-edge';
-
-    /** @var string Continuation of an entry that continues on the next grid row */
-    public const ACROSS_RIGHT_EDGE = 'across-right-edge';
-
-    /** @var string Continuation of an entry that started on a previous grid row and continues on the next */
-    public const ACROSS_BOTH_EDGES = 'across-both-edges';
-
-    protected $tag = 'div';
-
-    protected $defaultAttributes = ['class' => 'entry'];
-
-    protected $id;
-
+    /** @var ?string The description */
     protected $description;
-
-    protected $start;
-
-    protected $end;
-
-    /** @var ?int The 0-based position of the row where to place this entry on the grid */
-    protected $position;
-
-    /** @var ?ContinuationType */
-    protected $continuationType;
-
-    protected $rrule;
-
-    /** @var Url */
-    protected $url;
-
-    protected $isOccurrence = false;
 
     /** @var Attendee */
     protected $attendee;
 
-    public function __construct(int $id)
-    {
-        $this->id = $id;
-    }
-
-    public function getId(): int
-    {
-        return $this->id;
-    }
-
+    /**
+     * Set the description
+     *
+     * @param ?string $description
+     *
+     * @return $this
+     */
     public function setDescription(?string $description): self
     {
         $this->description = $description;
@@ -88,119 +36,23 @@ class Entry extends BaseHtmlElement
         return $this;
     }
 
+    /**
+     * Get the description
+     *
+     * @return ?string
+     */
     public function getDescription(): ?string
     {
         return $this->description;
     }
 
-    public function setStart(DateTime $start): self
-    {
-        $this->start = $start;
-
-        return $this;
-    }
-
-    public function getStart(): ?DateTime
-    {
-        return $this->start;
-    }
-
-    public function setEnd(DateTime $end): self
-    {
-        $this->end = $end;
-
-        return $this;
-    }
-
-    public function getEnd(): ?DateTime
-    {
-        return $this->end;
-    }
-
     /**
-     * Set the position of the row where to place this entry on the grid
+     * Set the attendee
      *
-     * @param ?int $position The 0-based position of the row
+     * @param Attendee $attendee
      *
      * @return $this
      */
-    public function setPosition(?int $position): self
-    {
-        $this->position = $position;
-
-        return $this;
-    }
-
-    /**
-     * Get the position of the row where to place this entry on the grid
-     *
-     * @return ?int The 0-based position of the row
-     */
-    public function getPosition(): ?int
-    {
-        return $this->position;
-    }
-
-    /**
-     * Set the continuation type of this entry
-     *
-     * @param ?ContinuationType $continuationType
-     *
-     * @return $this
-     */
-    public function setContinuationType(?string $continuationType): self
-    {
-        $this->continuationType = $continuationType;
-
-        return $this;
-    }
-
-    /**
-     * Get the continuation type of this entry
-     *
-     * @return ?ContinuationType
-     */
-    public function getContinuationType(): ?string
-    {
-        return $this->continuationType;
-    }
-
-    public function setRecurrencyRule(?string $rrule): self
-    {
-        $this->rrule = $rrule;
-
-        return $this;
-    }
-
-    public function getRecurrencyRule(): ?string
-    {
-        return $this->rrule;
-    }
-
-    public function setIsOccurrence(bool $state = true): self
-    {
-        $this->isOccurrence = $state;
-
-        return $this;
-    }
-
-    public function isOccurrence(): bool
-    {
-        return $this->isOccurrence;
-    }
-
-    public function setUrl(?Url $url): self
-    {
-        $this->url = $url;
-
-        return $this;
-    }
-
-    public function getUrl(): ?Url
-    {
-        return $this->url;
-    }
-
     public function setAttendee(Attendee $attendee): self
     {
         $this->attendee = $attendee;
@@ -208,34 +60,23 @@ class Entry extends BaseHtmlElement
         return $this;
     }
 
+    /**
+     * Get the attendee
+     *
+     * @return Attendee
+     */
     public function getAttendee(): Attendee
     {
         return $this->attendee;
     }
 
-    protected function assemble()
+    public function getColor(int $transparency): string
     {
-        $this->getAttributes()
-            ->add('data-entry-id', $this->getId())
-            ->add('data-entry-position', $this->getPosition());
+        return TimeGrid\Util::calculateEntryColor($this->getAttendee()->getName(), $transparency);
+    }
 
-        $continuationType = $this->getContinuationType();
-        if ($continuationType === self::ACROSS_GRID || $continuationType === self::ACROSS_BOTH_EDGES) {
-            $this->getAttributes()->add('class', 'two-way-gradient');
-        } elseif ($continuationType === self::FROM_PREV_GRID || $continuationType === self::ACROSS_LEFT_EDGE) {
-            $this->getAttributes()->add('class', 'opening-gradient');
-        } elseif ($continuationType === self::TO_NEXT_GRID || $continuationType === self::ACROSS_RIGHT_EDGE) {
-            $this->getAttributes()->add('class', 'ending-gradient');
-        }
-
-        if (($url = $this->getUrl()) !== null) {
-            $entryContainer = new Link(null, $url);
-            $entryContainer->openInModal();
-            $this->addHtml($entryContainer);
-        } else {
-            $entryContainer = $this;
-        }
-
+    protected function assembleContainer(BaseHtmlElement $container): void
+    {
         $title = new HtmlElement('div', Attributes::create(['class' => 'title']));
         $content = new HtmlElement(
             'div',
@@ -250,6 +91,7 @@ class Entry extends BaseHtmlElement
         $startText = null;
         $endText = null;
 
+        $continuationType = $this->getContinuationType();
         if ($continuationType === self::ACROSS_GRID) {
             $startText = sprintf($this->translate('starts %s'), $this->getStart()->format('d/m/y'));
             $endText = sprintf($this->translate('ends %s'), $this->getEnd()->format('d/m/y H:i'));
@@ -311,6 +153,6 @@ class Entry extends BaseHtmlElement
             );
         }
 
-        $entryContainer->addHtml($content);
+        $container->addHtml($content);
     }
 }
